Have Open in Nuclide links check all open windows for projects
Summary:
Before prompting to open a previous project root, scan all the open windows to check if any of them contain it.

If so, forward the deep link to the other window :D
This shouldn't result in an infinite cycle, since the other window will immediately be able to open the file.

Reviewed By: jgebhardt

Differential Revision: D4778504

fbshipit-source-id: 1ed4b11a2568d88a375850efbc46bf3c02c5d87c