Properly handle '--' arguments
Summary:
This was an issue because for buck run '--' is required, but for debug builds we launch the executable directly, and no '--' is needed. We now automatically throw in the '--' if a platform provider supports run arguments, which also allows having '--' as an argument if the built binary depends on it.

Also fixes a bug where the arg arrays weren't being merged correctly if provided from both the settings box and the platform provider.

Reviewed By: hansonw

Differential Revision: D4795969

fbshipit-source-id: a87c9421f9c3c0bd68c18c6c762d994b5c958348