Ensure diagnostics are removed when the corresponding project root is removed
Summary:
One FlowProcess is disposed, this just signals consumers that there is not
currently a valid IDE connection, and then completes the stream. This has the
effect that everything gets completed and unsubscribed from as desired, and
`ensureInvalidations`, which is called down the line, makes sure that the
messages are cleared out.

Reviewed By: hansonw

Differential Revision: D4799077

fbshipit-source-id: ad5711d493ab602bfa87afa1c2d3441c26035437