Fix HHVM debugger request window spinner doesn't always dismiss
Summary: Sometimes the loading spinner in the requests window doesn't dismiss reliably while debugging PHP because the frontend is waiting for a threads updated notification from the back end. This change always sends an update message for the "enabled" (active, currently focused) thread when the debugger pauses, to ensure the UX updates properly.

Reviewed By: shushz

Differential Revision: D4801851

fbshipit-source-id: 8449fad73d468f8331c56c652be9704030401136