Fix "HHVM breakpoints still active after deleting" issue
Summary:
This issue has been coming up frequently in our support groups. The issue is that while we display different PHP requests as if they were different threads in the same program, they are not like that. Each request is independent, and they're actually serviced by different processes on the server.

When a breakpoint is in a common code path, multiple requests hit it independently. When you resume (F8) one request, we then context switch to the next request at the breakpoint and it appears as if that one just hit the bp.  If you remove a breakpoint while broken in to the debugger, but there are already other requests queueing up at that breakpoint, they will appear to continue hitting it after the breakpoint is removed. In reality, they hit it before it was removed, but this is not apparent from the UX in the frontend unless you carefully examine the Requests window. It's also possible they hit it shortly after the bp was removed, but there's a race between us telling HHVM and the request telling us it's broken in.

Even within the same request ID, there is a race between our debug backend telling HHVM the bp is removed and the next time the bp might hit on the same request if it happens to be on a "hot" path (like in my video below).

In this change, the Nuclide PHP debugger checks the breakpoint store on hitting a file-line breakpoint and verifies that the bp is still there. If it is not, it silently resumes the request rather than breaking to the front-end.

Reviewed By: shushz

Differential Revision: D4801097

fbshipit-source-id: 74f4d028ad54c44539a1749277d805333c422faf