Support workspace/symbol (symbol search) in LSP
Summary:
Adds support for project-wide symbol search.

For the icons, I picked ones close to the existing HackSymbolProvider icons where possible. In cases where none were applicable, I tried to pick icons similar to those that VSCode uses.

Note: Hack used to invite you to type special characters to configure the queries, like foo or %bar. Very few people used those characters, and it's not part of LSP. We could reconstruct the feature on the client-side if we wanted but it doesn't seem worth it. So I've removed the invitation.

Reviewed By: hansonw

Differential Revision: D4771981

fbshipit-source-id: 87b8b4fd78c4ff1d8d6bab20f15d2b04335eeb33