implement textDocument/findReferences
Summary:
This adds support for FindReferences in Nuclide's LSP.

Here is the corresponding Hack diff: D4783324

Here is a screenshot using Hack language service status quo:
{F66277367}

Here is a screenshot using LSP:
{F66277371}

Note that the status quo uses the fully-qualified symbol name in both the editor tab name (where it is far too long!) and in the title of the find-results window.

LSP doesn't offer up the symbol name at all, neither brief nor fully-qualified. I wanted to write Nuclide to work with any LSP client, so I wanted to see how far I could get by just with LSP as is. (Indeed, VSCode highlights the symbol with existing LSP servers). So on the Nuclide side, it derives the symbol name by using the contents of the text buffer at the specified range.

Reviewed By: hansonw

Differential Revision: D4783358

fbshipit-source-id: 56a28e949d08501c3dd427f3683c29997c45cd40