implement textDocument/typeCoverage
Summary:
I am adding a new request to LSP, called "typeCoverage", to deliver the Nuclide feature.

I wrote it to be universal enough to carry all the features of the current Hack, Flow and Java coverage providers. Basically, the response to this new textDocument/typeCoverage request is in 1:1 correspondence to what the Atom Language Service expects. So, all the processing that formerly was done by the HackTypeCoverageProvider now has to be done by the Hack language server itself.

Reviewed By: nmote

Differential Revision: D4798999

fbshipit-source-id: e8f2fef901882cd611c14a9f461b8b04e01a9d9b