Format-code picks tighter ranges to format
Summary:
Currently, if you have text like this

  the quick
  brown fox
  jumped over
  the lazy dog

and you select the middle two entire rows ending up with your caret at column 0 on the fourth row

  the quick
  **brown fox
  jumped over
  |the lazy dog

and you do Edit > Text > FormatCode, then Nuclide thinks you meant to format the last THREE lines.

I think that's wrong! You clearly only meant to format the middle two lines!

(This isn't an essential change, and it's not blocking anything, but it just bugs me as I'm testing code-formatting for LSP and want to get it out of the way...)

Reviewed By: hansonw

Differential Revision: D4803021

fbshipit-source-id: 78a04377205bd39d392addd513397ca31cfa07c7