make evaluationExpression support regex directly
Summary:
The editor needs to know what is the expression under the cursor, so it can send that expression to the debugger.

Nuclide in theory allows the language service to figure this out intelligently using the AST. But in practice it never actually did this: instead it does it all on the client by regex...

  Hack: /[a-zA-Z0-9_$]+/g
  Flow: /[$_a-zA-Z][$_\w]*/g
  All other languages: /\w+/gi

I've added the ability to specify the regex directly inside the AtomLanguageServiceConfig. This avoids having to make a round-trip to the server solely to use a custom regex. It also cleaned up + simplified some code.

For LSP purposes, it's also essential because LSP has no similar functionality at all. Therefore, in things like HackLanguage.js which configures its LSP service, it will provide the regexp to use (in addition to other configuration parameters like the path to the LSP server).

Reviewed By: peterhal

Differential Revision: D4811431

fbshipit-source-id: 75fc239ad50bec28bf57d589be52008743b0b57f