Have debugger frontend refresh scopes after evaluateOnCallFrame
Summary:
This moves the logic to refresh scopes after an eval() to the front-end. Since an evaluation can cause side effects, we need to refresh the visible variables after an evaluation so that they update in the UI. By having the front-end pull this info rather than the Java debug server try to push it, it's more reliable and this will pave the way to fix this for PHP and C++ debuggers as well, which have the exact same bug.

The PHP debugger doesn't currently support the getThreadStack operation, so this ends up being a no-op. I'll implement that and then PHP will work correctly too.

Reviewed By: shushz

Differential Revision: D4816678

fbshipit-source-id: 9a8ddeeea7640367151a4aceb003e9c5c6c62417