Migrate the nuclide-code-format range API to use TextEdits
Summary:
Instead of just replacing the target range, a code formatting provider could theoretically modify arbitrary parts of the file. This is necessary for LSP compatibility, mostly.

Since the existing behavior is a very specific case of the new behavior, it's a very straightforward conversion.

Note that we'll keep the `formatEntireFile` API around for providers which can't / don't provide text edits (in which case diffing the file is still necessary).

Reviewed By: ljw1004

Differential Revision: D4824806

fbshipit-source-id: 91f3fa3bdbad84b53654de304303d6da8977a5be