Add tooltip to data cells
Summary: If a table is width constrained, sometimes columns are too small to see all the text. You can of course resize the containing pane in most cases in Nuclide, but it'd be nice to just hover the mouse over a cell and see all the text.  Hit this need in the debugger pane where the call stack can be kind of squeezed.

Reviewed By: jgebhardt

Differential Revision: D4816672

fbshipit-source-id: bfbbc0496992b1b4a8b2085641f5829a3b5abb5a