NativeDebuggerService.getAttachTargetInfoList() is completely destroying Nuclide server connection
Summary:
This routine runs a `ps` command on the remote machine to get a list of native targets the debugger can be attached to. On my devserver, there are enough processes running that this takes a while to come back, and saturates the nuclide remote connection. This routine is queued periodically to refresh the UX while the launch attach window is open - but it's actually taking longer to come back than the timer, so the requests are piling up to the point where they run constantly and the nuclide server connection becomes unusable, breaking everything else in Nuclide.

This change ensures a new request is not issued if the old one hasn't come back yet.

Reviewed By: tjfryan

Differential Revision: D4830336

fbshipit-source-id: 7662b18793a416d7239babefb4ce7a363f9a13fc