PHP/Hack debugger tells user that the console REPL is ready for use way too soon
Summary:
When the PHP/Hack debugger first launches and connects to HHVM, one of the first things it does is open a new connection it calls the "dummy connection". This connection invokes a well-known PHP script called xdebug_include.php that the debugger requires to be installed at a predetermined location on the webserver.

xdebug_include.php has some PHP require() statements that cause it to include basically the entire website. It pulls into the local scope all of the globals, modules and routines that one can reference from the REPL, so that the variables and types all resolve. After that's done, it issues a hard break instruction to the runtime by calling xdebug_break(). This breaks this connection into the debugger, and serves as the stack frame call site for all REPL instructions that are issued on the console when the debugger is not stopped at a breakpoint on a specific connection.

require()'ing all that code takes some time, and until it's complete, this connection isn't ready to receive commands. Currently the code is causing us to tell the user the connection is ready as soon as the debugger sees the script start. In reality, it's not ready until the script finishes including things and issues a breakpoint.

This change fixes this by not issuing the message until the first breakpoint is seen from the dummy.

NOTE: Unfortunately this is going to make it look like preloading suddenly takes way longer than it used to. In reality, it was always taking this long and we were lying to the user about being ready.

Reviewed By: shushz

Differential Revision: D4836674

fbshipit-source-id: ddb632e22346e367e1f455b084c96cd91683ab10