Refactor adb into sdb and proper adb
Summary:
The code was very strange doing adb and sdb things with all the functions called adb.
I'm refactoring it by rename the package with the debug-bridge name and refactoring functions out
into three classed: DebugBridge, Adb and Sdb. This brings more clarity to the functions that are android or tizen only and
allows not duplicating a lot of checks.

Reviewed By: stepanhruda

Differential Revision: D4840023

fbshipit-source-id: 78ccfd478b2f35eaa6854d49cef3aee8878b73d3