Showing all error messages in console when running arcanist command
Summary:
A while ago, updating a revision would display a success notif even if arc failed (for lint, unit tests, other).  A fix was put into place to throw an error upon receiving the first error message, but that would hide later error messages that had further details.  This reverts the previous fix and fixes the issue without creating the side-effect bug.
My rxjs knowledge is a little bit weak so I'm not sure if concat was a good way to go about doing what I wanted to do.

Reviewed By: aadisriram

Differential Revision: D4839078

fbshipit-source-id: bfac7935c828eb42d29c518a840df199f0d8be10