Prevent needless calls to setState in FileTreeSidebarComponent.
Summary:
The FileTreeSidebarComponent gets re-rendered about once a second because of throttle(remeasureEvents)..._updateScrollerHeight(). Normally this wouldn't be a problem because React diffs the resultant HTML before applying it.

But... it doesn't play nice with Atom tooltips: if an Atom tooltip is up, it actually gets hidden when render() is next called. [I think atom only has an imperative not declarative way of doing tooltips, and therefore the declarative API for adding them "ref={addTooltip(...)}" is imperative under the hood, and therefore render() always has diffs. But I'm not sure].

Anyway, what I've done is refactor the throttle(..)._updateScrollerHeight. Now it gets an observable stream of scroller heights, and only sets state if the values in this stream change.

Reviewed By: jgebhardt

Differential Revision: D4739126

fbshipit-source-id: 86e6c8c189504462ea8b2cfdc02a2a9fcd1afcd3