Add gatekeeper for the feature "uncommitted/head/stack"
Summary:
I've added a gatekeeper for the new feature to show uncommitted/head/stack changes.

This gatekeeper doesn't remove the setting described in package.json (I can't see a way to do that).

Instead it (1) modifies FileTreeActions to ignore the setting and instead always obtain the uncommitted changes; (2) modifies FileTreeSidebarComponent to never display a fancy UI.

Note: I fetch the GK immediately when this package is loaded, rather than on-demand. That's because it's impossible to even render the UI until we know the GK.

Reviewed By: mostafaeweda

Differential Revision: D4739324

fbshipit-source-id: 38f509a9e86eddbfa99aad5351ad693f3a315b59