Use currentTarget rather than target to get editor
Summary:
When propagating the right-click event for source-control menu items, event.target is the original target clicked and event.currentTarget is the current parent container we are attempting to dispatch the event to.  Since we define the handlers on Atom's editor element, the currentTarget will always give us back the editor we want rather than some random div from the middle of the document.

https://github.com/facebook/nuclide/issues/1086

Reviewed By: mostafaeweda

Differential Revision: D4848784

fbshipit-source-id: d16475613058e8b605ef0a0ba47e9d2e855ba9be