Nuclide debugger UX - "Copy" is not offered in the context menu for primitive variables in the scopes pane
Summary: In the scopes pane, if you right click on a variable name that is a complex variable with members (which is rendered in an expander control), you get an option to copy the value to the clipboard. This isn't included for simple variables that don't get rendered in an expander. It should always be offered.

Reviewed By: jgebhardt

Differential Revision: D4893707

fbshipit-source-id: 5bede9ccc06fef11ddc0a46b6d82ade1a38dd93c