Fix exception in Java debugger when breaking in and no source can be found for any frame in the call stack
Summary: If source cannot be found for the top frame, currently the first frame with source is the one that is opened. However, if no frames can find source, a bogus URI is passed to the routine that opens a source file, which ends up giving it to nuclideUri, which throws an exception that is not caught. We should just silently fail to open source in this case (the call stack window already indicates the error with a big red triangle)

Reviewed By: shushz

Differential Revision: D4893534

fbshipit-source-id: 65872c2bc5da310526e66f15561ffe97bc752f09