Fix [Nuclide Debugger HHVM] uncaughtException:; name: Error; message: Cannot find xdebug breakpoint
Summary: When a breakpoint resolved notification is received from HHVM, we're throwing an error if the breakpoint can't be found in the store. This shouldn't be an error condition - breakpoint resolved is completely asynchronous, and HHVM might not resolve the bp until the script is loaded, which is at an arbitrary time in the future. By then the user could have legitimiately removed the breakpoint from the front-end, in which case it won't be in the store.

Reviewed By: mostafaeweda

Differential Revision: D4899600

fbshipit-source-id: 6f01787909770786e00d0908c973741479b3b57f