Fix test runner display issue
Summary:
If Atom's `PaneElement::showItemView()` is called with a visible element, it can clear the display value. Normally, this shouldn't happen, but it is in our case because workspace-views is adding to the pane directly. Since that functionality is upstreamed in 1.17, I think this fix makes the most sense.

Changelog: Fixed a bug whereby the contents of the test runner panel would be hidden.

Reviewed By: hansonw

Differential Revision: D4911851

fbshipit-source-id: 8c738d7f4e9b84c7929ffd5944d44aef91f24037