Create a store for different db paths
Summary:
This allows clients to register a certain adb path.
When the rpc wants to use any adb, it will try one by one starting the server until one works and then it'll use that one. In subsequent calls, it will give more priority
to the last path that worked.
I'll actually use this in my next diff

Reviewed By: sas

Differential Revision: D4914034

fbshipit-source-id: 4e064d29128c292c5c6fed69de8b38ac3c50bc73