Add FileTreeContextMenu.addToShowInSection
Summary:
This adds support to augment the "ShowIn" section of the Nuclide file tree context menu.  The main change required to support this is to put group separators at the maximum priority within the menu group so that new items show up before the separator.

Note that current "ShowIn" items are added in two places so they also show up in the core Atom context menu accessed from panes (see existing comment about "a bit of hackery going on here").  This change does not propagate that behavior to new items.  The method defined here is specifically about adding items to the FileTreeContextMenu.

This also adds a unit test for FileTreeContextMenu.

While adding the unit test, I noticed that `dispose()` was not actually removing items, so this also includes a fix for that.

Reviewed By: mostafaeweda

Differential Revision: D4924823

fbshipit-source-id: 8ff06c92adbe6adcf04d21d2703c0aa8de34af06