Show the available connections
Summary:
I've seen that there's no unified way to get notified when the list of current remote connections changes.
I added a function in the remote connections package that returns and observable that will emit changes.

I'm using it to display the list of available hosts on the device panel. I'll use the same function later to notify the different hosts of the different valid adb/sdb paths.

Reviewed By: stepanhruda

Differential Revision: D4920343

fbshipit-source-id: 1cf3de14af8eebcc304b2ac9eb14bbc15c187079