Nuclide Debugger: fix call stack line numbers off by one
Summary: The chrome protocol uses 0-based line numbers, we aren't correcting for that in the call stack component, causing the line numbers to be off by one. Note that when you click on the frame, it opens the text editor to the correct line, this is just wrong in the call stack pane.

Reviewed By: shushz

Differential Revision: D4925514

fbshipit-source-id: 8ffc74c89cbd27bb4e8232edbef9a10cc668f357