Introduced a new Atom command: nuclide-buck:open-nearest-build-file.
Summary:
Frequently I want to open the `BUCK` file that corresponds to the current file
that I am editing. This can be hard to find in the file tree if the file is in a
directory with a lot of entries. More commonly, it is not in the file tree at
all if I happened to open a file that is outside of my working set. Having a
special command for this eliminates the need for navigation in the file tree.

Reviewed By: mostafaeweda

Differential Revision: D4932669

fbshipit-source-id: 3bf4a4f704dd5f715f64b81ed41db878f8e6d589