Update babel & friends
Summary:
This update carries with it Babylon 6.17.0 which adds support for flow type spread, and babel-eslint 7.2.3, which has a fix for working with flow type spreads. The other upgrades are to dedupe the babel deps.

Unfortunately, the `no-alert` rule throws for some reason when there's a type spread (even on the latest ESLint). So I disabled it since it's not very important rule: "disallow the use of alert, confirm, and prompt". http://eslint.org/docs/rules/no-alert

Reviewed By: vjeux

Differential Revision: D4936744

fbshipit-source-id: b3a039f0b1508f26fdc1048268c29eaedaee42b7