Fix the log4js file appender shutdown function
Summary:
The log4js file appender shutdown function is currently wrong:
https://github.com/nomiddlename/log4js-node/blob/v1.1.1/lib/appenders/file.js#L117

In the Nuclide logs this causes something along the lines of:

```
log4js.fileAppender - Writing to file /tmp/nuclide-hansonw-logs/nuclide.log, error happened  { Error: write after end
    at writeAfterEnd (_stream_writable.js:166:12)
    at WriteStream.Writable.write (_stream_writable.js:217:5)
```

In particular, `write()` returning true does not necessarily mean it's safe to `end()` the stream. I've already merged in a patch to fix this but the next log4js is blocked on a lot of things.. so we can just patch it out.

Reviewed By: shushz

Differential Revision: D4937056

fbshipit-source-id: 3af5b2401ac5c976f82d4684f8d2897cbaea96e4