Introduce FileSystemService::findNearestAncestorNamed() and deprecate findNearestFile().
Summary:
In my recent change to add a command to "open the nearest build file" (D4932669),
I discovered that it did not work for remote files because `FileSystemService::findNearestFile()`
operated on `string` rather than `NuclideUri`. This meant that, even though this is a service
framework API, it would not automatically rewrite a remote `NuclideUri` as a local path before
invoking the method, which is what was causing the failure.

Apparently `FileSystemService` is pretty old code that predates the modern service framework
and has yet to be ported to it (which is why it uses `string`). Because there are a number of uses of
`FileSystemService::findNearestFile()` in existence, I did not want to change the API and audit
all of the existing call sites just to fix the "open the nearest build file" issue.
Instead, I decided to introduce a new API (which "open the nearest build file" now uses)
and mark the old one deprecated.

Also, the old API returns the path to the directory containing the file rather than the path to the file
itself. I found this unintuitive, so the new API I introduced returns the path to the file. As you
can see, this simplifies the code in `pkg/nuclide-buck/lib/buildFiles.js`.

Reviewed By: mostafaeweda

Differential Revision: D4935927

fbshipit-source-id: 7cc22801fa2a1e00ea23a312da345099824b0ae2